decoupling IteratorQuery into new simplified Traits
